requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
pydantic==2.11.7
pydantic-settings==2.11.0
PyYAML==6.0.2
httpx==0.28.1
ultralytics==8.3.187